requirements.txt
audioread==3.0.0
autopep8==1.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
Cython==3.0.0
decorator==5.1.1
distlib==0.3.6
einops==0.6.0
face-alignment==1.3.4
ffmpeg-python==0.2.0
filelock==3.8.0
fonttools==4.38.0
future==0.18.3
huggingface-hub==0.11.0
idna==3.4
imageio==2.31.1
joblib==1.3.2
kiwisolver==1.4.4
lazy_loader==0.3
librosa==0.9.1
llvmlite==0.40.1
matplotlib==3.6.2
mpmath==1.2.1
mutagen==1.46.0
networkx==3.0rc1
numba==0.57.1
numpy==1.23.5
opencv-python==4.8.0.76
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pooch==1.7.0
pycodestyle==2.9.1
pycparser==2.21
pydub==0.25.1
pynormalize==0.1.4
pyparsing==3.0.9
pysptk==0.2.1
python-dateutil==2.8.2
PyWavelets==1.4.1
pyworld==0.3.4
PyYAML==6.0
requests==2.28.1
resampy==0.4.2
Resemblyzer==0.1.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.9.3
six==1.16.0
soundfile==0.12.1
sympy==1.11.1
threadpoolctl==3.2.0
tifffile==2023.8.12
timm==0.6.11
toml==0.10.2
torch==1.14.0.dev20221120
torchaudio==0.14.0.dev20221120
torchvision==0.14.0
tqdm==4.64.1
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.6
webrtcvad==2.0.10
